pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-customer-support-ops"
version = "1.0.0"
description = "Real-world OpenEnv environment for customer support operations"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.0",
  "requests>=2.32.0",
  "httpx>=0.27.0",
  "PyYAML>=6.0.2"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0"
]

[project.scripts]
server = "server.app:main"
baseline = "baseline:main"
validate-submission = "scripts.validate_submission:main"

[tool.setuptools]
py-modules = ["models", "support_env", "baseline", "client"]
packages = ["server", "scripts"]
